improve Inventory performance with memoizing Item component
